Add semaphore for API queries to SQL
