+++ api/token.go
@@ -22,7 +22,7 @@ func (a *API) getToken(c *gin.Context) {
 	}
 	tokenID := common.TokenID(*tokenIDUint)
 	// Fetch token from historyDB
-	token, err := a.h.GetToken(tokenID)
+	token, err := a.h.GetTokenAPI(tokenID)
 	if err != nil {
 		retSQLErr(err, c)
 		return
@@ -45,7 +45,7 @@ func (a *API) getTokens(c *gin.Context) {
 		return
 	}
 	// Fetch exits from historyDB
-	tokens, pendingItems, err := a.h.GetTokens(
+	tokens, pendingItems, err := a.h.GetTokensAPI(
 		tokenIDs, symbols, name, fromItem, limit, order,
 	)
 	if err != nil {
